+++ packages/core/stories/components/typography/CounterController.tsx
@@ -1,11 +1,11 @@
 import { useState } from 'react'
-import { Chart, Surface, View, Counter, CounterProps, CounterValueProps } from '../../../src'
+import { Chart, Surface, View, Counter, CounterProps, TextContentProps } from '../../../src'
 
 const randomUniformValue = (min: number, max: number) => {
     return min + Math.random() * (max - min)
 }
 
-export const CustomCounterValue = ({ style, className, children }: CounterValueProps) => {
+export const CustomCounterValue = ({ style, className, children }: TextContentProps) => {
     return (
         <text style={style} className={className}>
             {children}